+++ Communication/Flash/FlashServer.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Plus.Communication.Abstractions;
 using Plus.Communication.Packets;
 
@@ -6,7 +7,7 @@ namespace Plus.Communication.Flash;
 
 public class FlashServer : TcpGameServer<FlashServerConfiguration>, IFlashServer
 {
-    public FlashServer(IOptions<FlashServerConfiguration> options, FlashClientFactory flashClientFactory, IPacketManager packetManager) : base(options, flashClientFactory, packetManager)
+    public FlashServer(IOptions<FlashServerConfiguration> options, FlashClientFactory flashClientFactory, IPacketManager packetManager, ILogger<FlashServer> logger) : base(options, flashClientFactory, packetManager, logger)
     {
     }
 }