+++ Communication/Nitro/NitroServer.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using NetCoreServer;
 using Plus.Communication.Abstractions;
 using Plus.Communication.Flash;
@@ -21,7 +22,7 @@ public interface INitroServer : IGameServer
 
 public class NitroServer : WebsocketGameServer<NitroServerConfiguration>, INitroServer
 {
-    public NitroServer(IOptions<NitroServerConfiguration> options, NitroClientFactory clientFactory, IPacketManager packetManager) : base(options, clientFactory, packetManager) { }
+    public NitroServer(IOptions<NitroServerConfiguration> options, NitroClientFactory clientFactory, IPacketManager packetManager, ILogger<NitroServer> logger) : base(options, clientFactory, packetManager, logger) { }
 }
 
 
@@ -29,18 +30,20 @@ public class NitroClientFactory : IGameClientFactory<WsSessionProxy, WsServer>
 {
     private readonly FlashPacketFactory _packetFactory;
     private readonly IRevisionsCache _revisionsCache;
+    private readonly ILogger<WsSessionProxy> _sessionLogger;
 
-    public NitroClientFactory(FlashPacketFactory packetFactory, IRevisionsCache revisionsCache)
+    public NitroClientFactory(FlashPacketFactory packetFactory, IRevisionsCache revisionsCache, ILogger<WsSessionProxy> sessionLogger)
     {
         _packetFactory = packetFactory;
         _revisionsCache = revisionsCache;
+        _sessionLogger = sessionLogger;
     }
 
     public WsSessionProxy Create(WsServer server)
     {
         var flashClient = new FlashGameClient((NitroServer)server, _packetFactory)
             { Revision = _revisionsCache.InternalRevision };
-        var wsSession = new WsSessionProxy((NitroServer)server, flashClient);
+        var wsSession = new WsSessionProxy((NitroServer)server, flashClient, _sessionLogger);
         return wsSession;
     }
 }